refactor(evm): compile templates out of the zkEVM build entirely
Replaces the constant flag with the build's own split: the recognizer, the
forwarder and dispatcher descriptors, and the frame-entry path all move to
*.std.cs, which the zkEVM build drops before compiling, and the two references
left in shared files become a partial method and an #if. Nothing about templates
survives into the guest — not the types, not the CodeInfo field, not a folded
branch — where before it relied on the compiler and ILC to remove a constant
false.

The types are partial so the analyzer accepts a suffixed file name; there is no
second part.

Verified against the built assemblies: CodeTemplate, MinimalProxy,
SelectorDispatch, EnterMinimalProxy and TryEnterFunctionBody all appear in the
standard Nethermind.Evm.dll and none appear in the zkEVM one.
